requirements.txt
fastapi
uvicorn
transformers
torch
langchain
cassio
streamlit
scikit-learn
pandas
matplotlib
sentence-transformers
tqdm
python-dotenv
hf_xet
langchain 
langchain-nvidia-ai-endpoints
langchain-community
python-dotenv
astrapy
newspaper3k
beautifulsoup4
requests
lxml_html_clean 
beautifulsoup4
streamlit_theme
feedparser